refactor #247: import utils from utils root folder
