+++ pages/api/player/[battleTag].ts
@@ -1,17 +1,17 @@
-import { calculateSaveDeathRatio } from '@/utils/calculateSaveDeathRatio'
-import { DetailedPlayerStats } from '@/interfaces/player'
-import { NextApiRequest, NextApiResponse } from 'next'
-import { formatRoundsData } from '@/utils/formatRoundsData'
-import { calculateTotals } from '@/utils/calculateTotals'
-import { blacklistedPlayers, roundNames } from '@/constants'
-import { formatGameAwards } from '@/utils/formatGameAwards'
-import { transformKeysToCamelCase } from '@/utils/transformKeysToCamelCase'
-import { calculateWinRate } from '@/utils/calculateWinRate'
-import { fetchData } from '@/utils/fetchData'
 import {
+  calculateSaveDeathRatio,
+  formatRoundsData,
+  calculateTotals,
+  formatGameAwards,
+  transformKeysToCamelCase,
+  calculateWinRate,
+  fetchData,
   calculateCompletedChallenges,
   calculateCompletedChallengesLegacy,
-} from '@/utils/calculateCompletedChallenges'
+} from '@/utils'
+import { DetailedPlayerStats } from '@/interfaces/player'
+import { NextApiRequest, NextApiResponse } from 'next'
+import { blacklistedPlayers, roundNames } from '@/constants'
 
 interface QueryParams {
   battleTag: string
